Most of the work is completed with the new Angular 5 UI

+++ src/main/java/com/dryadandnaiad/sethlans/domains/database/EntityObject.java
@@ -27,7 +27,7 @@ package com.dryadandnaiad.sethlans.domains.database;
  */
 public interface EntityObject {
 
-    Integer getId();
+    Long getId();
 
-    void setId(Integer id);
+    void setId(Long id);
 }
